version

+++ app/Http/Controllers/GroupController.php
@@ -12,13 +12,14 @@ use App\Acl\Eloquent\Group;
 
 use App\ActiveDirectory\Eloquent\Directory;
 
+use Sentinel;
 use Cartalyst\Sentinel\Users\EloquentUser;
 
 class GroupController extends Controller
 {
     public function __construct()
     {
-        $this->middleware('privilege:sys_admin', [ 'except' => [ 'search' ] ]);
+        $this->middleware('privilege:sys_admin', [ 'only' => [ 'index', 'delMultiGroups' ] ]);
         parent::__construct();
     }
 
@@ -34,10 +35,72 @@ class GroupController extends Controller
         if ($s)
         {
             $groups = Group::Where('name', 'like', '%' . $s .  '%')
+                ->where(function ($query) {
+                    $query->where('principal.id', $this->user->id)
+                        ->orWhere(function ($query) {
+                            $query->where('public_scope', '3')->where('users', $this->user->id);
+                        })
+                        ->orWhere(function ($query) {
+                            $query->where('public_scope', '<>', '2')->where('public_scope', '<>', '3');
+                        });
+                    })
                 ->get([ 'name' ]);
 
         }
+
         return Response()->json([ 'ecode' => 0, 'data' => $groups ]);
+    }
+
+    /**
+     * Display a listing of the resource.
+     *
+     * @return \Illuminate\Http\Response
+     */
+    public function mygroup(Request $request)
+    {
+        $query = Group::where('name', '<>', '');
+
+        if ($name = $request->input('name'))
+        {
+            $query->where('name', 'like', '%' . $name . '%');
+        }
+
+        if ($scale = $request->input('scale'))
+        {
+            if ($scale == 'myprincipal')
+            {
+                $query->where('principal.id', $this->user->id);
+            }
+            else if ($scale == 'myjoin')
+            {
+                $query->where('users', $this->user->id);
+            }
+        }
+
+        $query->where(function ($query) {
+            $query->where('principal.id', $this->user->id)
+                ->orWhere(function ($query) {
+                    $query->where('public_scope', '<>', '2')->where('users', $this->user->id);
+                });
+                //->orWhere(function ($query) {
+                //    $query->where('public_scope', '<>', '2')->where('public_scope', '<>', '3');
+                //});
+        });
+
+        // get total
+        $total = $query->count();
+
+        $page_size = 30;
+        $page = $request->input('page') ?: 1;
+        $query = $query->skip($page_size * ($page - 1))->take($page_size);
+        $groups = $query->get();
+
+        foreach ($groups as $group)
+        {
+            $group->users = EloquentUser::find($group->users ?: []);
+        }
+
+        return Response()->json([ 'ecode' => 0, 'data' => $groups, 'options' => [ 'total' => $total, 'sizePerPage' => $page_size ] ]);
     }
 
     /**
@@ -59,13 +122,26 @@ class GroupController extends Controller
             $query->where('directory', $directory);
         }
 
+        // public_scope 范围
+        if ($public_scope = $request->input('public_scope'))
+        {
+            if ($public_scope == '1')
+            {
+                $query->where('public_scope', '<>', '2')->where('public_scope', '<>', '3');
+            }
+            else if ($public_scope == '2' || $public_scope == '3')
+            {
+                $query->where('public_scope', $public_scope);
+            }
+        }
+
         // get total
         $total = $query->count();
 
         $page_size = 30;
         $page = $request->input('page') ?: 1;
         $query = $query->skip($page_size * ($page - 1))->take($page_size);
-        $groups = $query->get([ 'name', 'users', 'directory' ]);
+        $groups = $query->get();
 
         foreach ($groups as $group)
         {
@@ -83,13 +159,54 @@ class GroupController extends Controller
      */
     public function store(Request $request)
     {
+        $insValues = [];
+
         $name = $request->input('name');
         if (!$name)
         {
             throw new \UnexpectedValueException('the name can not be empty.', -10200);
         }
+        $insValues['name'] = $name;
 
-        $group = Group::create($request->all());
+        $principal = $request->input('principal');
+        if (isset($principal) && $principal)
+        {
+            if ($principal == 'self')
+            {
+                $insValues['principal'] = [ 'id' => $this->user->id, 'name' => $this->user->first_name, 'email' => $this->user->email ];
+            }
+            else
+            {
+                $principal_info = Sentinel::findById($principal);
+                if (!$principal_info)
+                {
+                    throw new \InvalidArgumentException('the user is not exists.', -10205);
+                }
+                $insValues['principal'] = [ 'id' => $principal_info->id, 'name' => $principal_info->first_name, 'email' => $principal_info->email ];
+            }
+        }
+
+        $public_scope = $request->input('public_scope'); 
+        if ($public_scope)
+        {
+            if (!in_array($public_scope, ['1', '2', '3']))
+            {
+                throw new \UnexpectedValueException('the public scope value has error.', -10204);
+            }
+            $insValues['public_scope'] = $public_scope;
+        }
+        else
+        {
+            $insValues['public_scope'] = '1';
+        }
+
+        $description = $request->input('description'); 
+        if (isset($description))
+        {
+            $insValues['description'] = $description;
+        }
+
+        $group = Group::create($insValues);
         return Response()->json([ 'ecode' => 0, 'data' => $group ]);
     }
 
@@ -120,6 +237,22 @@ class GroupController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $group = Group::find($id);
+        if (!$group)
+        {
+            throw new \UnexpectedValueException('the group does not exist.', -10201);
+        }
+
+        if (isset($group->diectory) && $group->directory && $group->diectory != 'self')
+        {
+            throw new \UnexpectedValueException('the group come from external directroy.', -10203);
+        }
+
+        if (!(isset($group->principal) && isset($group->principal['id']) && $group->principal['id'] == $this->user->id) && !$this->user->hasAccess('sys_admin'))
+        {
+            return Response()->json(['ecode' => -10002, 'emsg' => 'permission denied.']);
+        }
+
         $updValues = [];
         $name = $request->input('name');
         if (isset($name))
@@ -131,20 +264,44 @@ class GroupController extends Controller
             $updValues['name'] = $name;
         }
 
+        $principal = $request->input('principal');
+        if (isset($principal))
+        {
+            if ($principal)
+            {
+                $principal_info = Sentinel::findById($principal);
+                if (!$principal_info)
+                {
+                    throw new \InvalidArgumentException('the user is not exists.', -10205);
+                }
+                $updValues['principal'] = [ 'id' => $principal_info->id, 'name' => $principal_info->first_name, 'email' => $principal_info->email ];
+            }
+            else
+            {
+                $updValues['principal'] = '';
+            }
+        }
+
         $user_ids = $request->input('users');
         if (isset($user_ids))
         {
             $updValues['users'] = $user_ids ?: [];
         }
 
-        $group = Group::find($id);
-        if (!$group)
+        $public_scope = $request->input('public_scope');
+        if (isset($public_scope))
         {
-            throw new \UnexpectedValueException('the group does not exist.', -10201);
+            if (!in_array($public_scope, ['1', '2', '3']))
+            {
+                throw new \UnexpectedValueException('the public scope value has error.', -10204);
+            }
+            $updValues['public_scope'] = $public_scope;
         }
-        if (isset($group->diectory) && $group->directory && $group->diectory != 'self')
+
+        $description = $request->input('description'); 
+        if (isset($description))
         {
-            throw new \UnexpectedValueException('the group come from external directroy.', -10203);
+            $updValues['description'] = $description;
         }
 
         $group->fill($updValues)->save();
@@ -165,9 +322,15 @@ class GroupController extends Controller
         {
             throw new \UnexpectedValueException('the group does not exist.', -10201);
         }
+
         if (isset($group->diectory) && $group->directory && $group->diectory != 'self')
         {
             throw new \UnexpectedValueException('the group come from external directroy.', -10203);
+        }
+
+        if (!(isset($group->principal) && isset($group->principal['id']) && $group->principal['id'] == $this->user->id) && !$this->user->hasAccess('sys_admin'))
+        {
+            return Response()->json(['ecode' => -10002, 'emsg' => 'permission denied.']);
         }
 
         Group::destroy($id);
